requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
fairseq>=0.12.2

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
audiomentations>=0.30.0
pydub>=0.25.1

# Dataset handling
datasets>=2.14.0
huggingface-hub>=0.16.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Deep learning utilities
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
umap-learn>=0.5.3

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Web application
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
joblib>=1.3.0

# Development tools
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
